change width of logo
This is an attempt to solve #3.
It works up to a browser size of >250px (on Desktop and Mobile), now.
For smaller resolutions  the menu drops in the next line. This is an overall issue of the ppsn template.
Solution might be to set an overall minimal page width of 250px.